pyproject.toml
[tool.poetry]
name = "backend-user-service"
version = "0.1.0"
description = ""
authors = ["Aleksei Lepekha <aleksei.lepekha@praktika.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.0"
asyncpg = "^0.29"
fastapi = "^0.101.0"
greenlet = "^3.0.3"
gunicorn = "^21.2.0"
pydantic = "^1.10.14"
pycountry = "^22.3"
SQLAlchemy = "^1.4.52"
uvicorn = "^0.29.0"
boto3 = "^1.28.22"
praktika-shared-lib = { version = "^6", source = 'praktika' }
python-dotenv = "^1.0.0"
prometheus-fastapi-instrumentator = "^6.1.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.44.1"}
aws-lambda-powertools = "^2.36.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.3.0"
pytest-dotenv = "^0.5.2"
pytest-env = "^0.8.2"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py310']

# https://hadialqattan.github.io/pycln/#/
[tool.pycln]
all = true
exclude = ['src/db/migrations']

# https://flakeheaven.readthedocs.io/en/latest/index.html
[tool.flakeheaven]
format = "grouped"
max-line-length = 120
doctests = true
show-source = true
max-complexity = 10
exclude = ['src/db/migrations/*']

# https://github.com/DmytroLitvinov/awesome-flake8-extensions
[tool.flakeheaven.plugins]
flake8-print = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pyflakes = ["+*"]

[tool.flakeheaven.exceptions."**/__init__.py"]
pyflakes = ["-F401", "-F403"]

[[tool.poetry.source]]
name = "praktika"
url = "https://praktika-pypi-419985171955.d.codeartifact.us-east-1.amazonaws.com/pypi/praktika-pypi/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
